requirements.txt
Flask==3.0.3
Flask-CORS==4.0.1
PyPDF2==3.0.1
scikit-learn==1.5.0
python-docx==1.1.2